up/down hex input update
Because element.className returns the WHOLE string of all the classnames, it is possible that you may have an element with “multiple classnames” (which are separated with spaces in the string, for made-up example: if it has the class “HexInput” and “DifferentFont”, element.className would've give you “HexInput DifferentFont”, and checking if the ENTIRE STRING of an element if it contains a specific class “HexInput” using “==” is not a good idea). Using regex is probably the only way to check if a string contains such a word.