bugfix and while -> for
bugfix: why even need to apply the up/down incrementor when the default number input itself ALREADY have the up/down incrementor? This was erroneously being applied to type="number" on the input that was initalized to have a value of 9007199254740992. Also setselectionrange does not work with number input.

while->for: yeah, I'm changing my programming style.